requirements.txt
# Backend dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pillow==10.2.0
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
paho-mqtt==1.6.1
websocket-client==1.7.0
python-jose==3.3.0

# Training dependencies
scikit-learn==1.6.1
joblib==1.3.2
tqdm==4.66.1
pandas==2.2.0
datasets==2.18.0
seaborn==0.13.2
matplotlib==3.8.2

# Development dependencies
pytest==8.0.0
black==24.1.1
flake8==7.0.0
